Recalculate width change instantly
Reuse file data that is already read into a numpy array,
and only slice it later (which is a 0-copy operation).
Do this in the Image class, so that it works for every filter.

The image height is rounded down.  Another option would be to round
the height up, but it would require disabling initial zero-padding
in all the filters, and either unifying the subsequent one
(get_displayable), or removing it too, and moving it to the Image
resizing function.  Then we could keep a numpy array large enough to
hold all the resizes so far, in order to avoid excessive
(de)allocations.

Closes #43